Remove float_extras
[unstable, deprecated since 1.11.0]
