--- src/doc/unstable-book/src/library-features/float-extras.md
@@ -1,7 +1,0 @@
-# `float_extras`
-
-The tracking issue for this feature is: [#27752]
-
-[#27752]: https://github.com/rust-lang/rust/issues/27752
-
-------------------------
